--- giorgi_tskaroveli/templates/add.html
@@ -1,14 +1,0 @@
-{% extends "base.html" %}
-{% block content %}
-<div class="jummptron">
-    <h1>Add Method in View</h1>
-    <p>Add new item using the form below</p>
-    <form method="post">
-        {{ form.hidden_tag() }}
-        {{ form.name.label }} {{ form.name() }}
-        {{ form.param1.label }} {{ form.param1() }}
-        {{ form.param2.label }} {{ form.param2() }}
-        {{ form.submit() }}
-    </form>
-</div>
-{% endblock %}
