--- giorgi_tskaroveli/templates/base.html
@@ -1,28 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Title</title>
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.0-beta3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-eOJMYsd53ii+scO/bJGFsiCZc+5NDVN2yr8+0RDqr0Ql0h+rP48ckxlpbzKgwra6" crossorigin="anonymous">
-</head>
-<body>
-
-<nav class="navbar navbar-expand-lg navbar-dark bg-warning">
-    <div class="navbar-nav">
-        <a class="nav" href="{{ url_for('home') }}" > Home </a>
-        <a class="nav" href="{{ url_for('add_item') }}" > Add </a>
-        <a class="nav" href="{{ url_for('list_items') }}" > List </a>
-        <a class="nav" href="{{ url_for('delete_item') }}" > Delete </a>
-    </div>
-
-
-</nav>
-
-{% block content %}
-
-{% endblock %}
-
-
-
-</body>
-</html>
