--- giorgi_tskaroveli/templates/home.html
@@ -1,7 +1,0 @@
-{% extends "base.html" %}
-{% block content %}
-<div class="jummptron">
-    <h1>Hello User</h1>
-    <p>This is my first site</p>
-</div>
-{% endblock %}
